core: launch_config: use double precision for auto type parameters
Previously we used floats, but XmlRpc stores doubles, so we should
make use of the full precision.

+++ rosmon_core/src/launch/launch_config.cpp
@@ -514,7 +514,7 @@ static XmlRpc::XmlRpcValue autoXmlRpcValue(const std::string& fullValue)
 		try { return boost::lexical_cast<int>(fullValue); }
 		catch(boost::bad_lexical_cast&) {}
 
-		try { return boost::lexical_cast<float>(fullValue); }
+		try { return boost::lexical_cast<double>(fullValue); }
 		catch(boost::bad_lexical_cast&) {}
 
 		return fullValue;
